02_harvester/dbtoolset/requirements.txt
pymysql==1.1.0
psycopg2==2.9.9
pyodbc==5.0.1
pytest==7.4.3
PyYAML==6.0.1
